Removed obsolete code
Remove code to load CEDET, since it is now shipped with emacs 23.2.

Changed name of default-tab-width to tab-width.
Removed the c-hungry-delete-key and c-auto-newline variable.
Removed the variable european-calendar-style and used
calendar-date-style instead.
Removed the variables view-diary-entries-initially and mark-diary-entries-in-calendar
and used calendar-view-diary-initially-flag and
calendar-mark-diary-entries-flag.
Changed set-default-font to set-frame-font.
Removed the function rolando-scheme-send-whole-buffer.