Changed configuration for .emacs .gnus and .ercrc.el
On .emacs:
Added w3m-session to the w3m configuration.
Added w3m-yt-view that allows the view of youtube videos on w3m.
When entering gdb-mode, turn off yasnippet so that TAB doesn't conflict
with the gdb expansion.
Added a keybinding to c-mode-common to call the function
ff-find-other-file and to use tabs when editing files.
Added a couple of hooks to mail-mode-hook when using emacsclient to
write mutt's email.

On .gnus:
Corrected the name of the nnimap-authinfo-file folder.
Don't show the first article when entering a group.
Collapse the threads when entering a group.
Automatically jump to the next group.

On .erccl.el:
Changed charset from iso-8859-1 to utf-8 so it works with bitlbee.