Changes to .emacs
Wrap (require 'cl) in a with-no-warnings form, so that emacs doesn't
complain when compiling .emacs.

W3m now *does* load the tabs that it had open the last time

On a lisp buffer, <f7> evals the entire buffer (the same as with an
elisp buffer).

Emacs now saves the story of the minibuffer between emacs sessions,
using savehist-mode.

The key "a" is activated on the dired buffer.

Start using smex (search m-x).