Rename "-empty" to "-empty?" and make it work on infinite lists
It was pointed to be in https://github.com/magnars/dash.el/pull/91
that `length` doesn't work with infinite lists and that we should use
the builtin `null` to check if a list is empty.

This commit changes the name of "-empty" to both "-empty?" and
"-empty-p". Those last two are now considered an alias for `null`.